requirements.txt
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.38.5
botocore==1.38.5
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
s3transfer==0.12.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tldextract==5.2.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
